Move merge policies to datumaro/components/merge directory
