+++ datumaro/components/hl_ops.py
@@ -13,7 +13,7 @@ from datumaro.components.environment import Environment
 from datumaro.components.exporter import Exporter
 from datumaro.components.filter import XPathAnnotationsFilter, XPathDatasetFilter
 from datumaro.components.launcher import Launcher, ModelTransform
-from datumaro.components.operations import ExactMerge
+from datumaro.components.merge.exact_merge import ExactMerge
 from datumaro.components.transformer import Transform
 from datumaro.components.validator import TaskType, Validator
 from datumaro.util import parse_str_enum_value
