+++ datumaro/components/merge/__init__.py
@@ -9,15 +9,15 @@ DEFAULT_MERGE_POLICY = "exact"
 
 def get_merger(merge_policy: str = DEFAULT_MERGE_POLICY, *args, **kwargs) -> Merger:
     if merge_policy == "union":
-        from datumaro.components.operations import UnionMerge
+        from .union_merge import UnionMerge
 
         return UnionMerge(*args, **kwargs)
     elif merge_policy == "intersect":
-        from datumaro.components.operations import IntersectMerge
+        from .intersect_merge import IntersectMerge
 
         return IntersectMerge(*args, **kwargs)
     elif merge_policy == "exact":
-        from datumaro.components.operations import ExactMerge
+        from .exact_merge import ExactMerge
 
         return ExactMerge(*args, **kwargs)
