+++ tests/unit/test_dataset.py
@@ -49,7 +49,7 @@ from datumaro.components.filter import (
 from datumaro.components.importer import FailingImportErrorPolicy, ImportErrorPolicy
 from datumaro.components.launcher import Launcher
 from datumaro.components.media import Image, MediaElement, Video
-from datumaro.components.operations import IntersectMerge
+from datumaro.components.merge.intersect_merge import IntersectMerge
 from datumaro.components.progress_reporting import NullProgressReporter, ProgressReporter
 from datumaro.components.transformer import ItemTransform, Transform
 from datumaro.plugins.transforms import ProjectInfos
